requirements.txt
dj-database-url==0.4.2
Django==1.8
djangorestframework==3.6.2
gunicorn==19.7.1
psycopg2==2.6.2
pycrypto==2.6.1
Pygments==2.1.3
whitenoise==3.3.0
